+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Adds inbound and outbound TLS support for gRPC. See `gprc.Creds`,
+  `grpc.WithTransportCredentials`, and `grpc.Transport.NewDialer` for usage.
 - Added `peer/peerlist/v2` which differs from the original `peer/peerlist` by
   replacing the use of `api/peer.ListImplementation` with
   `peer/peerlist/v2.Implementation`, which threads the peer separately from the
@@ -31,8 +33,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change log level to reflect error statuses. Previously all logs were logged at
   debug level. Errors are now logged at error level.
 - Update pin for gogo/protobuf to ^1
-- Add `grpc.ClientTLSConfig` and `grpc.ServerTLSConfig` to permit the use of arbitrary
-  tls.Config structs with gRPC.
 
 ## [1.30.0] - 2018-05-03
 ### Added